package: move react-bootstrap, react, react-dom to peerDependencies, bump react-bootstrap (#79)
Fixes https://github.com/Imballinst/react-bs-datatable/issues/75 and fixes https://github.com/Imballinst/react-bs-datatable/issues/76. This PR fixes the bug where the library isn't compatible when the main app is using `react@^17`. On top of that, this library also bumps the main requirement for `react-bootstrap` so that the vulnerability issue is gone.