feat(responsive): add default mobile first stylings (#106)
Fixes https://github.com/imballinst/react-bs-datatable/issues/82.

Demo: https://codesandbox.io/s/mobile-first-2-2-0-k2g3l?file=/src/index.js.

This only adds the column sizing and proper responsive margins to the table control above. The rest is mostly off to the custom CSS. For example, in the sandbox above, this style is necessary to add:

```
.controlRow__root {
  margin-bottom: 15px;
}
```

So that there is a "distance" between the table control and the actual table.